% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Doptimize.R
\name{Defficiencies}
\alias{Defficiencies}
\title{Wrapper function for OApackage Defficiencies function.}
\usage{
Defficiencies(A)
}
\arguments{
\item{A}{An array containing the design (in 0, 1 format)}
}
\value{
A list containing the calculated efficiencies
}
\description{
This function calculates the D, Ds- and D1-efficiency of a design. The definitions
of these efficiencies can be found on the main page of the package \link{oapackage}.
}

