% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Doptimize.R
\name{Defficiencies}
\alias{Defficiencies}
\title{Wrapper function for OApackage Defficiencies function.}
\usage{
Defficiencies(N, k, A)
}
\arguments{
\item{N}{Number of runs}

\item{k}{Number of factors}

\item{A}{Array containing the design (in 0, 1 format)}
}
\value{
A list containing the calculated efficiencies
}
\description{
This function calculates the D, Ds- and D1-efficiency of a design.
}

